jinja-sql-runner-poc/requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyYAML==6.0.3
tabulate==0.9.0
